bugfix timeline; code cleanup
php script that returns list of images to display:
Now a serverside check is performed if the requested resource
exists and is only returned if it does so.
This fixes the issue of having blank frames being played back
in the image player.

Cleanup of unused stylesheet classes, and adjustment of nav-tab
that displays the loaded resources.