Addition of ModeHandler
The purpose of the ModeHandler is to deal with the state and
UI changes that are influenced by the currently selected mode.

Checkbox settings (which plots are currently viewed) are saved
when a mode switch occurs. The ref/curr <-> from/to textfield
is also left the same.

The previously known global variable global_mode is now
accessible through ModeHandler.getMode()